requirements.txt
cffi==0.9.2
-e git+git@github.com:django/django.git@feac4c30ce7635e17ce1adc4c2c7a1eb0721aeb3#egg=Django-dev
django-reversion==1.8.6
-e git+git@github.com:onespacemedia/cms.git@def9ee6fe10042d277ddd103a5d0fdcaf2d52bfc#egg=onespacemedia_cms-master
Pillow==2.8.1
Werkzeug==0.10.4
